// src/contracts/generated/csi07/RecoveryActionV1.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type RecoveryActionV1 = "retrySameOperation" | "retryCredentials" | "enterCredentials" | "useExistingAccount" | "restartWithRequiredAssurance" | "restartProductAuth" | "reauthenticate" | "refreshCatalog" | "none";