// src/contracts/generated/csi07/ProtocolErrorDetailsV1.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CatalogVersion } from "./CatalogVersion";
import type { PasswordFieldV1 } from "./PasswordFieldV1";
import type { PasswordViolationV1 } from "./PasswordViolationV1";

export type ProtocolErrorDetailsV1 = { "kind": "empty" } | { "kind": "flowLimit", limit: number, } | { "kind": "catalogStale", expectedCatalogVersion: CatalogVersion, } | { "kind": "rateLimit", retryAfterSeconds: number, } | { "kind": "passwordPolicy", field: PasswordFieldV1, violations: Array<PasswordViolationV1>, } | { "kind": "callbackOperation", callbackOperationId: string, };